Moved definition of DAC class from __init__.py to proper file.
