+++ qctoolkit/hardware/dacs/__init__.py
@@ -1,26 +1,6 @@
-from abc import ABCMeta, abstractmethod
-from typing import Dict, Tuple
+from qctoolkit.hardware.dacs.dac_base import *
 
-import numpy
-
-__all__ = ['DAC']
-
-
-class DAC(metaclass=ABCMeta):
-    """Representation of a data acquisition card"""
-
-    @abstractmethod
-    def register_measurement_windows(self, program_name: str, windows: Dict[str, Tuple['numpy.ndarray',
-                                                                                       'numpy.ndarray']]) -> None:
-        """"""
-
-    @abstractmethod
-    def register_operations(self, program_name: str, operations) -> None:
-        """"""
-
-    @abstractmethod
-    def arm_program(self, program_name: str) -> None:
-        """"""
-
-    def delete_program(self, program_name) -> None:
-        """"""
+try:
+    from qctoolkit.hardware.dacs.alazar import *
+except ImportError:
+    pass
